wip: fixing imports on refactored packages
