+++ non_linear/binary_search_tree/binary_search_tree.go
@@ -1,4 +1,4 @@
-package non_linear
+package binary_search_tree
 
 import (
 	"fmt"
