+++ non_linear/counter/counter.go
